requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
Django==4.0.10
django-cors-headers==3.13.0
django-formtools==2.4
django-otp==1.1.4
django-phonenumber-field==6.4.0
django-two-factor-auth==1.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==1.7.1
pytz==2022.7.1
qrcode==7.3.1
requests==2.28.2
six==1.16.0
slugify==0.0.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
